controlling the io and parser threads from the learner thread

+++ vowpalwabbit/parse_example.h
@@ -26,6 +26,6 @@ void read_lines(vw* all, const char* line, size_t len,
 }  // namespace VW
 
 int read_features_string(vw* all, v_array<example*>& examples, std::vector<VW::string_view>& words, std::vector<VW::string_view>& parse_name);
-size_t read_features(vw *all, std::vector<char>& line, size_t& num_chars, v_array<example*>& examples);
+int read_features(vw *all, std::vector<char>& line, size_t& num_chars, v_array<example*>& examples);
 size_t strip_features_string(char*& line, size_t num_chars_init);
